pyproject.toml
[tool.poetry]
name = "chat_ai"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^2.7"
#python3="^3.6"
dialogflow = "^0.7.2"
apiai = "^1.2.3"

[tool.poetry.dev-dependencies]
pytest = "^4.6"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
